+++ Core/Matcher/LocationMatcher.php
@@ -20,9 +20,11 @@ class LocationMatcher extends RepositoryMatcher
     const MATCH_LOCATION_REMOTE_ID = 'location_remote_id';
     const MATCH_PARENT_LOCATION_ID = 'parent_location_id';
     const MATCH_PARENT_LOCATION_REMOTE_ID = 'parent_location_remote_id';
+    const MATCH_CONTENT_TYPE_ID = 'contenttype_id';
+    const MATCH_CONTENT_TYPE_IDENTIFIER = 'contenttype_identifier';
 
     protected $allowedConditions = array(
-        self::MATCH_AND, self::MATCH_OR,
+        self::MATCH_AND, self::MATCH_OR, self::MATCH_NOT,
         self::MATCH_CONTENT_ID, self::MATCH_LOCATION_ID, self::MATCH_CONTENT_REMOTE_ID, self::MATCH_LOCATION_REMOTE_ID,
         self::MATCH_PARENT_LOCATION_ID, self::MATCH_PARENT_LOCATION_REMOTE_ID
     );
@@ -46,6 +48,21 @@ class LocationMatcher extends RepositoryMatcher
         $this->validateConditions($conditions);
 
         foreach ($conditions as $key => $values) {
+
+            $query = new LocationQuery();
+            $query->limit = PHP_INT_MAX;
+            $query->filter = $this->getQueryCriterion($key, $values);
+            $results = $this->repository->getSearchService()->findLocations($query);
+
+            $locations = [];
+            foreach ($results->searchHits as $result) {
+                $locations[$result->valueObject->id] = $result->valueObject;
+            }
+
+            return new LocationCollection($locations);
+        }
+
+        /*foreach ($conditions as $key => $values) {
 
             if (!is_array($values)) {
                 $values = array($values);
@@ -76,7 +93,7 @@ class LocationMatcher extends RepositoryMatcher
                 case self::MATCH_OR:
                     return $this->matchOr($values);
             }
-        }
+        }*/
     }
 
     /**
@@ -97,6 +114,7 @@ class LocationMatcher extends RepositoryMatcher
      *
      * @param int[] $contentIds
      * @return Location[]
+     * @deprecated
      */
     protected function findLocationsByContentIds(array $contentIds)
     {
@@ -115,6 +133,7 @@ class LocationMatcher extends RepositoryMatcher
      *
      * @param int[] $remoteContentIds
      * @return Location[]
+     * @deprecated
      */
     protected function findLocationsByContentRemoteIds(array $remoteContentIds)
     {
@@ -131,6 +150,7 @@ class LocationMatcher extends RepositoryMatcher
     /**
      * @param int[] $locationIds
      * @return Location[]
+     * @deprecated
      */
     protected function findLocationsByLocationIds(array $locationIds)
     {
@@ -146,6 +166,7 @@ class LocationMatcher extends RepositoryMatcher
     /**
      * @param int[] $locationRemoteIds
      * @return Location[]
+     * @deprecated
      */
     protected function findLocationsByLocationRemoteIds($locationRemoteIds)
     {
@@ -161,6 +182,7 @@ class LocationMatcher extends RepositoryMatcher
     /**
      * @param int[] $parentLocationIds
      * @return Location[]
+     * @deprecated
      */
     protected function findLocationsByParentLocationIds($parentLocationIds)
     {
@@ -182,6 +204,7 @@ class LocationMatcher extends RepositoryMatcher
     /**
      * @param int[] $parentLocationRemoteIds
      * @return Location[]
+     * @deprecated
      */
     protected function findLocationsByParentLocationRemoteIds($parentLocationRemoteIds)
     {
@@ -193,5 +216,79 @@ class LocationMatcher extends RepositoryMatcher
         }
 
         return $this->findLocationsByParentLocationIds($locationIds);
+    }
+
+    /**
+     * @param $key
+     * @param $values
+     * @return mixed should it be \eZ\Publish\API\Repository\Values\Content\Query\CriterionInterface ?
+     * @throws \Exception for unsupported keys
+     */
+    protected function getQueryCriterion($key, $values)
+    {
+        if (!is_array($values)) {
+            $values = array($values);
+        }
+
+        switch ($key) {
+            case self::MATCH_CONTENT_ID:
+                return new Query\Criterion\ContentId($values);
+
+            case self::MATCH_LOCATION_ID:
+                return new Query\Criterion\LocationId($values);
+
+            case self::MATCH_CONTENT_REMOTE_ID:
+                return new Query\Criterion\RemoteId(reset($values));
+
+            case self::MATCH_LOCATION_REMOTE_ID:
+                return new Query\Criterion\LocationRemoteId($values);
+
+            case self::MATCH_PARENT_LOCATION_ID:
+                return new Query\Criterion\ParentLocationId($values);
+
+            case self::MATCH_PARENT_LOCATION_REMOTE_ID:
+                $locationIds = [];
+                foreach ($values as $remoteParentLocationId) {
+                    $location = $this->repository->getLocationService()->loadLocationByRemoteId($remoteParentLocationId);
+                    // unique locations
+                    $locationIds[$location->id] = $location->id;
+                }
+                return new Query\Criterion\ParentLocationId($locationIds);
+
+            case 'content_type_id':
+            case self::MATCH_CONTENT_TYPE_ID:
+                return new Query\Criterion\ContentTypeId($values);
+
+            case 'content_type_identifier':
+            case self::MATCH_CONTENT_TYPE_IDENTIFIER:
+                return new Query\Criterion\ContentTypeIdentifier($values);
+
+            case self::MATCH_AND:
+                $subCriteria = array();
+                foreach($values as $subCriterion) {
+                    $value = reset($subCriterion);
+                    $subCriteria[] = $this->getQueryCriterion(key($subCriterion), $value);
+                }
+                return new Query\Criterion\LogicalAnd($subCriteria);
+
+            case self::MATCH_OR:
+                $subCriteria = array();
+                foreach($values as $subCriterion) {
+                    $value = reset($subCriterion);
+                    $subCriteria[] = $this->getQueryCriterion(key($subCriterion), $value);
+                }
+                return new Query\Criterion\LogicalOr($subCriteria);
+
+            case self::MATCH_NOT:
+                /// @todo throw if more than one sub-criteria found
+                $subCriterion = reset($values);
+                $value = reset($subCriterion);
+                $subCriterion = $this->getQueryCriterion(key($subCriterion), $value);
+                return new Query\Criterion\LogicalNot($subCriterion);
+
+            default:
+                throw new \Exception($this->returns . " can not be matched because matching condition '$key' is not supported. Supported conditions are: " .
+                    implode(', ', $this->allowedConditions));
+        }
     }
 }